requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
greenlet==3.0.3
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==5.3.0
Mako==1.3.0
MarkupSafe==2.1.3
orjson==3.9.10
passlib==1.7.4
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
pypng==0.20220715.0
PyQRCode==1.2.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
qrcode==7.4.2
qrcode-styled==0.2.2
requests==2.31.0
requests-async==0.6.2
rfc3986==1.5.0
rsa==4.9
setuptools==75.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.24
sse==1.2
sse-starlette==1.8.2
starlette==0.32.0.post1
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.25.0
watchfiles==0.21.0
websockets==12.0
